requirements.txt
pandas
rdkit-pypi
numpy
sklearn